% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alf_http.R
\name{add_params}
\alias{add_params}
\title{Add parameters}
\usage{
add_params(endpoint, params, sep = "?")
}
\arguments{
\item{endpoint}{endpoint URI to add parameters to}

\item{params}{list of name/value parameters}

\item{sep}{separator used to delimit the next parameter to the endpoint}
}
\value{
endpoint with added parameters
}
\description{
Add parameters to endpoint
}
