% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alf_node.R
\name{alf_node.new}
\alias{alf_node.new}
\title{Create a new Alfresco node}
\usage{
alf_node.new(session, node_id, node_details)
}
\arguments{
\item{session}{valid Alfresco repository session}

\item{node_id}{node id}

\item{node_details}{details of new node}
}
\value{
node details
}
\description{
Creates a new Alfresco repository node as a child of \code{node_id}.
}
