% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alf_ticket.R
\name{alf_ticket}
\alias{alf_ticket}
\title{Get Alfresco authentication ticket}
\usage{
alf_ticket(uri, username, password)
}
\arguments{
\item{uri}{base uri to Alfresco repository}

\item{username}{user name}

\item{password}{password}
}
\value{
Alfresco ticket that can be used to authenticate subsequent calls to the repository
}
\description{
Get Alfresco authentication ticket for specified respository with provided credentials.
}
