% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfticket.R
\name{alfticket}
\alias{alfticket}
\title{Get Alfresco authentication ticket}
\usage{
alfticket(uri, username, password)
}
\arguments{
\item{uri}{base uri to Alfresco repository}

\item{username}{user name}

\item{password}{password}
}
\value{
Alfresco ticket that can be used to authenticate subsequent calls to the repository
}
\description{
Get Alfresco authentication ticket for specified respository with provided credentials.
}
