% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggexport.R
\name{ggexport}
\alias{ggexport}
\title{Simpler ggsave with useful defaults}
\usage{
ggexport(filename, size = "medium", dpi = 96, font = NULL,
  extension = "pdf")
}
\arguments{
\item{filename}{Destination path of the exported plot}

\item{size}{Selected plot size, to choose between "medium" and "small" or as a vector}

\item{dpi}{DPI used when exporting the plot}

\item{font}{Optional selected font, to choose between "lmodern" and "palatino"}

\item{extension}{Extension of the exported plot, either "pdf" or "png"}
}
\description{
Alternative implementation of ggsave (from ggplot2 package) with nice defaults. Only GNU/Linux with libnotify macOS are supported right now.
}
\details{
The function does the following things:

\itemize{
  \item execute any R call
  \item print the elapsed time to execute it
  \item send a permanent desktop notification
  \item emit a completion alert sound
}
}
\examples{

}
\keyword{system.time}
