% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lok_regar.R
\name{lok_regar}
\alias{lok_regar}
\title{Ready for Orders (Lok-Regar)}
\usage{
lok_regar(call)
}
\arguments{
\item{call}{Any (?) R call}
}
\description{
This function is used to control execution time when running long simulations. Only GNU/Linux with libnotify macOS are supported right now.
}
\details{
The function does the following things:

\itemize{
  \item execute any R call
  \item print the elapsed time to execute it
  \item send a permanent desktop notification
  \item emit a completion alert sound
}
}
\examples{
lok_regar(print("hello world"))
lok_regar(a <- seq(1:10000))
}
\keyword{system.time}
