% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alf_http.R
\name{alf_GET}
\alias{alf_GET}
\title{TODO}
\usage{
alf_GET(endpoint, ticket, params = list())
}
\arguments{
\item{endpoint}{TODO}

\item{ticket}{TODO}

\item{params}{TODO}
}
\value{
TODO
}
\description{
TODO
}
