% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alf_http.R
\name{alf_method}
\alias{alf_method}
\title{Alfresco HTTP method}
\usage{
alf_method(method = c("GET", "POST"), endpoint, ticket,
  params = list(), format = c("json", "file"), body = NULL)
}
\description{
Alfresco HTTP method
}
