% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alf_node.R
\name{alf_node.new}
\alias{alf_node.new}
\title{TODO}
\usage{
alf_node.new(session, node_id, node_details)
}
\arguments{
\item{session}{TODO}

\item{node_id}{TODO}

\item{node_details}{TODO}
}
\value{
TODO
}
\description{
TODO
}
