% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alf_node.R
\name{alf_node}
\alias{alf_node}
\title{Get Alfresco node}
\usage{
alf_node(session, node_id = "-root-", relative_path = NULL)
}
\arguments{
\item{session}{valid Alfresco repository session}

\item{node_id}{node id, defaults to \code{-root-}}

\item{relative_path}{relative path from \code{node_id} to required node, defaults to \code{NULL}}
}
\value{
Node details
}
\description{
Gets the details of an Alfresco repository node matching \code{node_id} or, if provided, the node at \code{relative_path}
relative to \code{node_id}.
}
\examples{
# try to establish a connection to the alfresco content repository
my_session <- tryCatch(
                alf_session("http://localhost:8080", "admin", "admin"),
                error = function(e) NULL)

if (!is.null(my_session)) {

  # generate unique document name
  my_document_name <- paste(round(proc.time()[[3]]), ".txt", sep="")
  relative_path <- paste("test/", my_document_name, sep="")

  # create document
  my_new_document <- alf_node.new(my_session, node_id="-root-",
    list(
      name = my_document_name,
      nodeType = "cm:content",
      relativePath = "test"
    ))

  # upload content
  my_new_document$content$update("R/examples/sample_content/sample.txt")

  # get details of document node
  my_document <- alf_node(my_session, relative_path = relative_path)

  # output the name of the document
  print(my_document$name)

  # output the details of documents content
  print(my_document$content$mime_type)
  print(my_document$content$mime_type_name)
  print(my_document$content$size)
  print(my_document$content$encoding)

  # read document content
  my_content_file <- file(my_document$content$as.file(), "r")
  my_content <- readLines(my_content_file)
  close(my_content_file)
  print(my_content)

  # upload new content
  my_updated_document <- my_document$content$update("R/examples/sample_content/modified_sample.txt")

  # print updated content size
  print(my_updated_document$content$size)
}
}
