% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alf_http.R
\name{alf_POST}
\alias{alf_POST}
\title{Alfresco POST method}
\usage{
alf_POST(endpoint, ticket, params = list(), format = c("json", "file"),
  body = NULL)
}
\arguments{
\item{endpoint}{base endpoint URI}

\item{ticket}{authentication ticket}

\item{params}{optional list of parameters}

\item{format}{return format from \code{json} (default)}

\item{body}{request body}
}
\value{
result based on format provided
}
\description{
Helper to make a POST call to the Alfresco REST API
}
