% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alf_session.R
\name{alf_session.invalidate}
\alias{alf_session.invalidate}
\title{Invalidates a session.}
\usage{
alf_session.invalidate(session)
}
\arguments{
\item{session}{session}
}
\description{
Invalidates a valid session so it can no longer be used to connect to an Alfresco repository.
}
\examples{
# try to establish a connection to the alfresco content repository
my_session <- tryCatch(
                alf_session("http://localhost:8080", "admin", "admin"),
                error = function(e) NULL)

if (!is.null(my_session)) {

  # output the server URL
  print(my_session$server)

  # output the connection ticket
  print(my_session$ticket)
}
}
