% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alf_session.R
\name{alf_session}
\alias{alf_session}
\title{Get connection session to Alfresco repository}
\usage{
alf_session(server, username, password)
}
\arguments{
\item{server}{Alfresco server URL}

\item{username}{user name}

\item{password}{password}
}
\value{
connection session to Alfresco repository
}
\description{
Get connection session to Alfresco repository including:
TODO list things available
}
