% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfr.R
\docType{package}
\name{alfr}
\alias{alfr}
\alias{alfr-package}
\title{alfr: A package for connecting with Alfresco}
\description{
The \code{alfr} package provides a way to connect to Alfresco and interact with the contents of the repository.
\itemize{
  \item \code{\link{alfticket}} - get Alfresco authentication ticket
  \item \code{\link{alfnode}} -
}
}
\author{
Roy Wetherall \email{rwetherall@gmail.com}
}
